backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
groq
pydantic[email]
python-dotenv
motor
pymongo
python-multipart
python-jose[cryptography]
passlib
bcrypt==4.0.1
email-validator
